Prevent scan_id clobbering
The server generates a scan id, checks it against elastic, and then sends that with the work request. This prevents agents from the unlikely chance they generate the same ID for two different scans and reduces likelihood of having multiple results for a single scan id in elastic.gethost_scan_id. Closes #49