Server controls ports that agents scan, but in a good way
This closes #40. Instead of creating a database table for individual ports and port ranges, I've created a versioned field for an nmap-services file. This nmap-services file is fetched by the agent over the api anytime it notices a change in the sha256 of the contents. Admins can upload this as a file in a new /admin/services view. This closes a huge issue for the natlas project, as we can now, finally, have the server tell agents what ports to scan - and in a reasonable fashion at that.