Upgrade agent documentation, improve screenshot reliability
The setup script for the agent was pretty awful, so this has improved that. We were also trying to use the requests library but never installing it. Setup agent was updated to try to install and set up the virtualenv just like the server. vnc screenshots now use xvfb-run, which I confirmed working on the natlas.io test server. Closes #28.