Run agent as root because of nmap features
TCP/IP fingerprinting requires elevated privileges. For this early version (and because it's already the normal behavior), exit if we're not running as root. Also added an error dictionary that maps to return codes and more elegantly handle the unexpected event of the data files we're trying to read not being found. This closes #85.