Increase nginx max body size to 10 Megabytes
The default nginx max body size is 1 megabyte. This is fine for smaller scans, but when scans return a lot of data (or when we start adding more data for scans to return), this 1 megabyte limit can become a gating factor and will cause the agents to get stuck in an infinite exponential backoff. This closes #93.