Improved nmap parser
This closes #69, we're no longer using grungy substring searches against the nmap data, and we're instead parsing the xml data to pull out information about the host. This gives us access to structured data about each port and we're storing data about whether the host is up or not. This sets us up nicely to log scan attempts for hosts that are deemed to be down, as well.