Configurable maximum retries on submission
The backoff request function accepts a boolean parameter to determine whether or not it should give up after the config defined max retries. This is only used for submissions because in certain cases submissions might be triggering internal server errors, either because of elastic or because the results are too big. Obviously losing the data isn't ideal, but neither is sitting in a locked state where the thread continually submits data that returns a 500. This will close #102.