Set SECRET_KEY during setup-server.sh
This will see if a SECRET_KEY is set in .env, and if it's not set then it'll check if it's set in the current environment. If it is, it'll copy that into the .env file. If it's not, it will generate a random 32 character secret key using /dev/urandom. This closes #101.