Move invite/reset tokens to session to prevent referrer leakage
This commit adds an intermediary step between clicking the link in your email and actually using the token for the password form. By doing this, we are able to hide the tokens from external referrer headers, as we aren't loading any external resources with the token in the url. This should close #139.