Automatically attempt to reconnect to elastic every 60 seconds
If we try to run an elasticsearch function and we're currently in a bad state, we will try to reconnect. If it's been less than 60 seconds since our last attempt, we immediately bail out. If we hit a connection exception during elastic functions then we'll set our status to false so that we can try to reconnect later. Closes #154.