Check for update button + Clickable version string
Check for update will check against the github api to see if there is a new latest tag and offer a link if there is. The version string in the footer can be clicked to open a new tab taking you to the github page about that particular version release. Closes #48.