Add screenshots to mapping and make ports an array instead of nested
We originally chose nested because I might have wanted to query a port individually. But I don't think that's ever actually the case. I will still want the rest of the information about the host. If I want to be able to view data about a specific port only, I can do that by extracting that port from the array. This opens up a bunch of better search options that don't require writing a nested query. This only applies to new elastic indices, so you'll have to migrate data if you want to make use of it.