Make the ssl parser more robust by checking if fields exist in source data before trying to copy them
Unfortunately ssl certificates are fickle creatures and we can't rely on every field to be present. For the sake of fixing exceptions, I made this very verbose change to ensure that we don't try to add any keys that don't exist. I think it can be refactored but I don't know how.