Functionality capable of clustering screenshots captured by natlas during scans.
Strategy is to support perceptual hash algorithms, and to use kmeans (after PCA reduction) to create clusters.

Clusters will always be ordered based on their norm from the all-0 (origin) vector. Basically, clusters are output in the order of their "complexity".

Three perceptual hashes are supported. The first, based on ImageMagick, performs the best, but requires that ImageMagick is installed. It also takes some significant time to run.

The next two are based on image scaling and DCT transforms, and are implemented by an open source python library. The open source version is directly used and a modified version taking all color channels is implemented as the final option.