Update README.md
After consulting with a lawyer, it's important to include the original Intel Corporation copyright notice to acknowledge the code contributed by Intel Corporation at the start of this project. I've also updated the main notice to reflect my github name as opposed to a useless "The Natlas Authors" entity which doesn't exist and doesn't actually assign copyright to anyone. I will be looking into creating a legal entity for the natlas project and then assigning the copyright to that entity.