Introduce execute_query pattern to enable easier instrumentation in the future
There is still work left to be done, I've only really implemented search and count right now. We also need to implement delete_by_query, index, and probably a few others. This commit also introduces a dedicated "browse" endpoint that is separate from the search endpoint. While similar in function, the search endpoint was getting unruly to manage both basic browse plus search. This allows us to simplify a bit both in the code as well as in the templates. There are definitely some abstractions left to do to simplify this further, though. Such as url generation for pagination, which is relatively similar all over the place and can be abstracted a little bit.