Update dockerfile, remove setuptools from listed dependencies
Manually installing a version of setuptools like that can cause issues if that version of setuptools isn't compatible with the installed version of pip. Updating to the python:3.7.5 means we've got an up to date pip and easyinstall version which work fine for the necessary dependencies. This base image also comes with libsqlite which means we don't have to apt-get update and ship our apt cache with the final image like we were previously doing.