1. Fixes a bug with the CyclicPRNG that caused it to be biased for N=2 The fix here is to include the N=2 case in the init_generator method and modify the method to use the range (2, N-1) rather than (2, N-2) for generator search. The range limited by N-2 was previously used as a small optimization for the search, since N-2 can only be a generator for when the cyclic group is of order 2.
2. Simplifies determining whether an address is within the scope and not in the blacklist
The IPScopeManager now keeps track of the scope and blacklist, and its methods use an O(log(log(n))) search instead of the O(log(n)) loops previously used.

3. Extends address validity for the above calculation to IPv6
Move to using the netaddr module for validation. We get IPv6 support for free from the move to netaddr module.

4. Moves all of the scope related code into one portion of the repository
/app/scope/ is the new directory. It has its own __init__ and looks like a separate module.

5. Refactor code to improve code quality and standardization
Move all scope code to snake_case for methods, and the rest of the code base to use the new method names. Handle some codeclimate and LGTM warnings, as well as those pointed out in PR review