Docker - Don't specify base image minor versions
Semver claims that minor/build version bumps are backwards
compatible. Looking at NodeJS's changelog, we're missing out on
a few vuln/bug fixes so we'll switch to major releases.

Python seems to make more breaking changes in an X.Y change, so we'll
stick with the major-minor versioning.

For NodeJS, we'll use X.*
For Python, we'll use X.Y.*.