Change: Automatic migration detection prevents app from launching if migration needed
This change closes #255 and introduces a new config option, DB_AUTO_MIGRATE, to provide optional automatic migration support. This is a change in the default behavior of previous versions, which would do a db upgrade at the start of every run even if unnecessary. By only doing migrate when we need to, it saves us an entire app init cycle, and providing the option to not migrate will support multi-node server deployments better.