Cache images we send from s3 (#1249)
Since all images are going to be served via the flask application now, we should set up some caching so that the browser doesn't download duplicates. We set the timeout to one day, which should be somewhat reasonable to balance the cache size and the server load. In most use cases, you probably won't need to reload the same images more than a day later, or if you do, it'll be in smaller views rather than the homepage that shows potentially hundreds of images in one page load.