Revamp the work API internals (#1250)
The `getwork` API currently does some annoying dict building to
determine what to return. This replaces that process with a pydantic
model that we dump to json at the moment we return the response,
ensuring type information is available for as long as possible within
the server.

This also switches from using the `current_app` in-memory values to just
querying the database for a specific record for the services and agent
config. While not my favorite, it's not that big a deal and it means
that it's one less thing in the web worker memory to get out of sync
when we have multiple workers running.

There's definitely some improvements to be made, in part in the
underlying database models, in part using enums instead of magic values
everywhere, and in part improving the ability to just automatically
serialize my models without having to do this weird `**Model.as_dict()`
pattern into a serializer constructor.

But as of right now, `/api/getwork` returns the exact same content as
before this PR, just in a less annoying way.