Change elastic interface calling convention (#1251)
Attaching things to the app instance makes mypy mad, and makes the DX
bad.

Instead, treat the elastic interface similar to how we treat db, scope
manager, login, etc. Which is that the process gets one instance that
just gets imported everywhere. This is functionally no different, I
think, than attaching one instance to app and then only accessing it
through `current_app`.

The elastic types themselves are still messy because I am dealing with a
lot of dicts. But we'll get there.