Update NatlasServices and AgentConfig (#1252)
NatlasServices and AgentConfig both had some remnants of
`current_app.<cached_value>` pattern, which we want to eliminate.

Additionally, they both had fields that were nullable previously that I
didn't know were nullable. None of them should be nullable, so migrate
them to not-null.

Finally, NatlasServices is being changed to update-in-place instead of
each new port generating a new entry in the table. This mimics the
AgentConfig behavior. We didn't expose the old versions of the
NatlasServices file anyways, so it was just an append-only table. We may
as well update it in place and always just be able to get 1.

One thing I don't love about AgentConfig and NatlasServices access
patterns is that I constantly have to deal with "what if it's None". I
think I am going to just add special functions to those two models that
automatically get pk=1 or automatically create and then get pk=1 if it
doesn't already exist. But that will be a separate PR.