Switch agent to using uv (#1253)
Pipfile has long been a pain in my side. It's slow, to say the least.

This PR updates the agent to use [uv](https://docs.astral.sh/uv/)
instead, which is considerably faster and more robust. Dependencies are
also stored in the standard pyproject.toml, and the lock file `uv.lock`
is better than `Pipfile.lock`, in my opinion.

I will also be moving the server, but it made sense to do one component
at a time.