Tailwind Transformation (#1258)
While Bootstrap has been great to me over the years, Natlas just feels
kinda dated. I want to go for a more modern feel. To that end, I've
updated basically every frontend file to use tailwind instead of
bootstrap. I've replaced much of bootstrap's javascript functionality
with Alpine.js, which I'm on the fence about. I like locality of
behavior a lot, but there are definitely a few design decisions that
I've made with Natlas over the years that make Alpine feel like you need
to hack it to get it to work.

Either way, this is basically a completely frontend rewrite, meant to
keep the _general_ feel of most of the application, but swapping up the
navigation, login, and a few other things.

This also further solidifies the lack of backwards compatibility, since
it removes the old html components for old natlas versions before
0.6.12. I'm going to also get rid of 0.6.12, but only after I rewrite
the data model to be versioned so that I can skip tying data versions to
application versions.