Move scripts to an array field on AgentConfig (#1274)
The AgentScript table exists solely as a list of script names. It wasn't
even related to the AgentConfig table or anything. Completely
unnecessary, I can just use an array field.

This migration loses state of the configured scripts in the AgentScript
table. But since the next release is going to be hard-incompatible, I'm
not worried about this at this time.